Move LayoutFile and LayoutFileEqualityComparer to R7.Epsilon.Models
