+++ R7.Epsilon.LayoutManager/Components/LayoutController.cs
@@ -25,10 +25,9 @@ using System.Diagnostics.Contracts;
 using System.IO;
 using System.Linq;
 using DotNetNuke.Common;
-using DotNetNuke.Data;
 using DotNetNuke.Entities.Portals;
 using R7.Epsilon.Components;
-using R7.Epsilon.LayoutManager.Models;
+using R7.Epsilon.Models;
 
 namespace R7.Epsilon.LayoutManager.Components
 {
