+++ R7.Epsilon/R7.Epsilon.csproj
@@ -217,6 +217,8 @@
     </Compile>
     <Compile Include="Components\Const.cs" />
     <Compile Include="Components\LayoutHelper.cs" />
+    <Compile Include="Models\LayoutFile.cs" />
+    <Compile Include="Models\LayoutFileEqualityComparer.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Containers\Blank.ascx" />
